# Conflicts:
#	Assets/Scenes/Scena Marco.unity

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,8 +1,9 @@
+using System;
 using UnityEngine;
 
 public interface IDamageable
 {
-	public int HP { get; set; }
+	public float HP { get; set; }
 	public Transform colliderTransform { get; set; }
 
 	public void TakeDamage(int damage);
